+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 ### Data Pack & Resource Pack Creators
 - Added support for MoLang conditions for quirks and poses.
 - Changed the AttackDefenceRatio requirement to StatCompare and StatEqual.
+- Added parametric motion and rotation support to particle effects.
 
 ## [1.5.0 - The Ruins and Revival Update (May 12th, 2024)](#1-5-0)
 #### "You're telling me that Mojang has added archaeology to the game? Hmm... that gives me an idea."